Simplify scope graph
