+++ crates/eww/src/main.rs
@@ -39,7 +39,6 @@ pub mod opts;
 pub mod script_var_handler;
 pub mod server;
 pub mod util;
-pub mod pettree;
 pub mod widgets;
 
 fn main() {
