+++ crates/eww/src/new_state_stuff.rs
@@ -1,13 +1,13 @@
 use std::{collections::HashMap, sync::Arc};
 
-use crate::pettree;
 use anyhow::*;
 use eww_shared_util::VarName;
-use petgraph::graph::NodeIndex;
+use petgraph::{
+    graph::{DiGraph, NodeIndex},
+    EdgeDirection::{Incoming, Outgoing},
+};
 use simplexpr::{dynval::DynVal, SimplExpr};
 use yuck::config::{widget_definition::WidgetDefinition, widget_use::WidgetUse, window_definition::WindowDefinition};
-
-pub type ScopeTree = pettree::ScopeTree<Scope>;
 
 pub fn do_stuff(
     global_vars: HashMap<VarName, DynVal>,
@@ -49,37 +49,10 @@ pub struct Scope {
 }
 
 impl Scope {
-    pub fn new(data: HashMap<VarName, DynVal>) -> Self {
+    /// Initializes a scope **incompletely**. The [`node_index`] is not set correctly, and needs to be
+    /// set to the index of the node in the scope graph that connects to this scope.
+    fn new(data: HashMap<VarName, DynVal>) -> Self {
         Self { data, listeners: HashMap::new(), node_index: NodeIndex::default() }
-    }
-
-    pub fn contains(&self, k: &VarName) -> bool {
-        self.data.contains_key(k)
-    }
-
-    pub fn get(&self, k: &VarName) -> Option<&DynVal> {
-        self.data.get(k)
-    }
-
-    fn register(&mut self, listener: Listener) -> Result<()> {
-        let listener = Arc::new(listener);
-        for needed_var in listener.needed_variables.iter() {
-            self.listeners.entry(needed_var.clone()).or_default().push(listener.clone());
-        }
-        Ok(())
-    }
-
-    fn update_value(&mut self, var_name: &VarName, value: DynVal) -> Result<()> {
-        if let Some(map_entry) = self.data.get_mut(var_name) {
-            *map_entry = value;
-        }
-        Ok(())
-    }
-}
-
-impl pettree::HasScopeContents for Scope {
-    fn has_variable(&self, var_name: &VarName) -> bool {
-        self.contains(var_name)
     }
 }
 
@@ -96,22 +69,45 @@ impl std::fmt::Debug for Listener {
 #[derive(Debug, Eq, PartialEq, Clone, Copy, Hash)]
 pub struct ListenerId(usize);
 
-impl pettree::ScopeTree<Scope> {
+#[derive(Debug, Eq, PartialEq)]
+enum ScopeTreeEdge {
+    ChildOf,
+    ReferencesVariable(VarName),
+}
+
+impl ScopeTreeEdge {
+    fn references_var(&self, name: &VarName) -> bool {
+        match self {
+            ScopeTreeEdge::ChildOf => false,
+            ScopeTreeEdge::ReferencesVariable(x) => x == name,
+        }
+    }
+}
+
+#[derive(Debug)]
+pub struct ScopeTree {
+    graph: DiGraph<Scope, ScopeTreeEdge>,
+    pub root_index: NodeIndex,
+}
+
+impl ScopeTree {
     pub fn from_global_vars(vars: HashMap<VarName, DynVal>) -> Self {
-        let mut tree = Self::new(Scope::new(vars));
-        let root_index = tree.root_index;
-        tree.value_at_mut(root_index).map(|scope| {
+        let mut graph = DiGraph::new();
+        let root_index = graph.add_node(Scope { data: vars, listeners: HashMap::new(), node_index: NodeIndex::default() });
+        graph.node_weight_mut(root_index).map(|scope| {
             scope.node_index = root_index;
         });
-        tree
+        Self { graph, root_index }
     }
 
-    pub fn add_scope(&mut self, child_of: NodeIndex, value: Scope) -> NodeIndex {
-        let node_index = self.add_node(child_of, value);
-        self.value_at_mut(node_index).map(|scope| {
-            scope.node_index = node_index;
+    pub fn add_scope(&mut self, child_of: NodeIndex, scope_variables: HashMap<VarName, DynVal>) -> NodeIndex {
+        let scope = Scope::new(scope_variables);
+        let new_index = self.graph.add_node(scope);
+        self.graph.add_edge(child_of, new_index, ScopeTreeEdge::ChildOf);
+        self.value_at_mut(new_index).map(|scope| {
+            scope.node_index = new_index;
         });
-        node_index
+        new_index
     }
 
     pub fn run_listeners_for_value_change(&mut self, index: NodeIndex, var_name: &VarName) -> Result<()> {
@@ -138,7 +134,11 @@ impl pettree::ScopeTree<Scope> {
 
     pub fn update_value(&mut self, index: NodeIndex, var_name: &VarName, value: DynVal) -> Result<()> {
         let index = self.find_scope_with_variable(index, var_name).context("Variable not found in scope")?;
-        self.value_at_mut(index).map(|scope| scope.update_value(var_name, value));
+        self.value_at_mut(index).map(|scope| {
+            if let Some(map_entry) = scope.data.get_mut(var_name) {
+                *map_entry = value;
+            }
+        });
         self.run_listeners_for_value_change(index, var_name)?;
 
         for child in self.children_referencing(index, var_name) {
@@ -149,21 +149,108 @@ impl pettree::ScopeTree<Scope> {
     }
 
     pub fn register_listener(&mut self, index: NodeIndex, listener: Listener) -> Result<()> {
+        // Set up the graph edges describing that a scope has a listener that references a variable from another scope.
         for needed_var in listener.needed_variables.iter() {
-            self.add_var_reference_to_node(index, needed_var.clone())?;
+            let scope = self.value_at(index).context("Given index is not in the graph")?;
+            if !scope.data.contains_key(needed_var) {
+                let mut cur_idx = index;
+                while let Some(parent) = self.parent_of(cur_idx) {
+                    let parent_scope = self.value_at(parent).expect("Nodes parent was not in the graph...");
+                    if parent_scope.data.contains_key(needed_var) {
+                        self.graph.add_edge(parent, index, ScopeTreeEdge::ReferencesVariable(needed_var.clone()));
+                        break;
+                    }
+                    cur_idx = parent;
+                }
+            }
         }
-        self.value_at_mut(index).map(|scope| scope.register(listener));
+        self.value_at_mut(index).map(|scope| {
+            let listener = Arc::new(listener);
+            for needed_var in listener.needed_variables.iter() {
+                scope.listeners.entry(needed_var.clone()).or_default().push(listener.clone());
+            }
+        });
         Ok(())
     }
 
     pub fn find_scope_with_variable(&self, index: NodeIndex, var_name: &VarName) -> Option<NodeIndex> {
-        self.find_ancestor_or_self(index, |scope| scope.contains(var_name))
+        self.find_ancestor_or_self(index, |scope| scope.data.contains_key(var_name))
     }
 
     pub fn lookup_variable_in_scope(&self, index: NodeIndex, var_name: &VarName) -> Option<&DynVal> {
         self.find_scope_with_variable(index, var_name)
             .and_then(|scope_index| self.value_at(scope_index))
-            .map(|x| x.get(var_name).unwrap())
+            .map(|x| x.data.get(var_name).unwrap())
+    }
+
+    pub fn remove_node_recursively(&mut self, index: NodeIndex) {
+        let mut children = self.graph.neighbors_directed(index, Outgoing).detach();
+        while let Some(child) = children.next_node(&self.graph) {
+            self.remove_node_recursively(child);
+        }
+        self.graph.remove_node(index);
+    }
+
+    pub fn value_at(&self, index: NodeIndex) -> Option<&Scope> {
+        self.graph.node_weight(index)
+    }
+
+    pub fn value_at_mut(&mut self, index: NodeIndex) -> Option<&mut Scope> {
+        self.graph.node_weight_mut(index)
+    }
+
+    pub fn parent_of(&self, index: NodeIndex) -> Option<NodeIndex> {
+        self.find_neighbor(index, Incoming, |edge| edge == &ScopeTreeEdge::ChildOf)
+    }
+
+    pub fn children_referencing(&self, index: NodeIndex, var_name: &VarName) -> Vec<NodeIndex> {
+        self.neighbors_where(index, Outgoing, |edge| edge.references_var(var_name))
+    }
+
+    fn find_neighbor(
+        &self,
+        index: NodeIndex,
+        dir: petgraph::EdgeDirection,
+        f: impl Fn(&ScopeTreeEdge) -> bool,
+    ) -> Option<NodeIndex> {
+        let mut neighbors = self.graph.neighbors_directed(index, dir).detach();
+        while let Some(neighbor) = neighbors.next_node(&self.graph) {
+            let edges = match dir {
+                Outgoing => self.graph.edges_connecting(index, neighbor),
+                Incoming => self.graph.edges_connecting(neighbor, index),
+            };
+            if edges.into_iter().any(|x| f(x.weight())) {
+                return Some(neighbor);
+            }
+        }
+        None
+    }
+
+    fn neighbors_where(
+        &self,
+        index: NodeIndex,
+        dir: petgraph::EdgeDirection,
+        f: impl Fn(&ScopeTreeEdge) -> bool,
+    ) -> Vec<NodeIndex> {
+        let mut neighbors = self.graph.neighbors_directed(index, dir).detach();
+        let mut result = Vec::new();
+        while let Some(neighbor) = neighbors.next_node(&self.graph) {
+            if self.graph.edges_connecting(index, neighbor).into_iter().any(|x| f(x.weight())) {
+                result.push(neighbor);
+            }
+        }
+        result
+    }
+
+    /// Search through the ancestors of a node for a value that satisfies the given predicate.
+    /// Also looks at the given node itself.
+    pub fn find_ancestor_or_self(&self, index: NodeIndex, f: impl Fn(&Scope) -> bool) -> Option<NodeIndex> {
+        let content = self.value_at(index)?;
+        if f(content) {
+            Some(index)
+        } else {
+            self.find_ancestor_or_self(self.parent_of(index)?, f)
+        }
     }
 }
 
@@ -183,9 +270,9 @@ mod test {
 
         let child_index = scope_tree.add_scope(
             scope_tree.root_index,
-            Scope::new(hashmap! {
+            hashmap! {
                 VarName("bar".to_string()) => DynVal::from("ho"),
-            }),
+            },
         );
 
         scope_tree
